% Generated by roxygen2 (4.0.1): do not edit by hand
\name{addToPram}
\alias{addToPram}
\title{addToPram}
\usage{
addToPram(pram, name, val)
}
\arguments{
\item{pram}{- Null or the existing list of parameters}

\item{name}{- the name of the sub-parameter list to add the value to}

\item{val}{- the value to add the the parameter ist.}
}
\description{
Description:
	Used to create and add to a list, parameters using the same name are added
to a list of other parameters using the same name. The list is then returned.
}

