% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od.R
\name{od}
\alias{od}
\title{Create origin-destination flows (odf) object}
\usage{
od(E, U = NULL, col_orig = 1, col_dest = 2, col_id = 1)
}
\arguments{
\item{E}{`data.frame` or `sf` object of edges. If it is an `sf`, it should contain (MULTI)LINESTRING geometries}

\item{U}{`sf` object of nodes. It should contain `POINT` geometry.}

\item{col_orig}{column of E (integer or name) that specifies the origin}

\item{col_dest}{column of E (integer or name) that specifies the destination}

\item{col_id}{column of U (integer or name) that specifies the id}
}
\description{
Create origin-destination flows (odf) object, which is a list consisting of a data.frame with the od data, an sf object with spatial points and optionally, an sf object with spatial (poly)lines of the routes
}
