% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_add_lines.R
\name{od_add_lines}
\alias{od_add_lines}
\title{Add lines (curved or straight) to an odf object}
\usage{
od_add_lines(
  x,
  angle = 1/24 * pi,
  points_per_line = ifelse(angle == 0, 2, 100),
  range = c(0, 1),
  trunc = units::set_units(c(0, 0), "m"),
  min_trunc_dist = units::set_units(5000, "m"),
  overwrite.geometry = FALSE
)
}
\arguments{
\item{x}{od object}

\item{angle}{angle}

\item{points_per_line}{points_per_line}

\item{range}{range}

\item{trunc}{trunc}

\item{min_trunc_dist}{= min_trunc_dist}

\item{overwrite.geometry}{overwrite.geometry}
}
\description{
Add lines (curved or straight) to an odf object
}
