% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_is_valid.R
\name{od_is_valid}
\alias{od_is_valid}
\title{Checck validity of od objects}
\usage{
od_is_valid(x, verbose = TRUE)
}
\arguments{
\item{x}{od object}

\item{verbose}{should messages be returned about possible invalidations? TRUE by default.}
}
\description{
Checks whether an od object is valid. See details.
}
\details{
A valid `od` object should: contain two list items, `U` and `E`, and three attributes `od_id`, `od_orig`, and `od_dest`. The values of these attributes are column names of `U` (`od_id`) and `E` (`od_orig` and `od_dest`). `U` should be an `sf` object (points or polygons), and `E` a data.frame (optionally an sf object (lines).
}
