% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_sum.R
\name{od_sum_in}
\alias{od_sum_in}
\alias{od_sum_out}
\alias{od_sum_stay}
\title{Sum of OD flows}
\usage{
od_sum_in(x, value, name = paste(value, "in", sep = "_"))

od_sum_out(x, value, name = paste(value, "out", sep = "_"))

od_sum_stay(x, value, name = paste(value, "stay", sep = "_"))
}
\arguments{
\item{x}{\code{od} object}

\item{value}{names of the variable in x$E that contain the values to be summed}

\item{name}{name of the newly created variable of sums in x$U}
}
\description{
Sum of OD flows.
}
