% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odf_add_lines.R
\name{odf_add_lines}
\alias{odf_add_lines}
\title{Add lines (curved or straight) to an odf object}
\usage{
odf_add_lines(x, via = TRUE, angle = 1/24 * pi, crs = 3857,
  points_per_line = 100)
}
\arguments{
\item{angle}{angle}

\item{crs}{crs}

\item{points_per_line}{points_per_line}

\item{odf}{odf object}
}
\description{
Add lines (curved or straight) to an odf object
}
