% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odf_map.R
\name{odf_flows}
\alias{odf_flows}
\alias{od_flows}
\alias{odf_points}
\title{Create map from odf object}
\usage{
od_flows(x, ...)

odf_points(x)
}
\arguments{
\item{x}{odf object}

\item{...}{arguments passed on to od_add_lines}
}
\description{
Create map from odf object
}
