% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odf_heatmap.R
\name{odf_heatmap}
\alias{odf_heatmap}
\title{Create heatmap}
\usage{
odf_heatmap(x, log = FALSE, n = 5, legend.title = "Flow")
}
\arguments{
\item{x}{odf object}

\item{log}{logarithmic scale?}

\item{n}{number of classes (if !log)}

\item{legend.title}{legend title}
}
\description{
Create heatmap
}
