% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odf_processing.R
\name{odf_remove_type}
\alias{odf_remove_type}
\title{Remove type grouping and aggregate odf accordingly}
\usage{
odf_remove_type(x, na.rm = FALSE)
}
\arguments{
\item{x}{odf object}

\item{na.rm}{na.rm}
}
\description{
Remove type grouping and aggregate odf accordingly
}
