% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odf_processing.R
\name{odf_remove_via}
\alias{odf_remove_via}
\title{Remove via points and aggregate odf accordingly}
\usage{
odf_remove_via(x, na.rm = FALSE)
}
\arguments{
\item{x}{odf object}

\item{na.rm}{na.rm}
}
\description{
Remove via points and aggregate odf accordingly
}
