% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{data}
\name{NL_commuting}
\alias{NL_commuting}
\alias{NL_muni_point}
\alias{NL_muni_poly}
\title{Dutch commuting data}
\usage{
data(NL_commuting)

data(NL_muni_point)

data(NL_muni_poly)
}
\description{
Dutch commuting data
}
\details{
\code{NL_commuting} Dutch commuting data from 2016 or 2017

\code{NL_muni_point} Spatial points of Dutch municipalities

\code{NL_muni_poly} Spatial polygons of Dutch municipalities
}
