% Generated by roxygen2 (4.0.1): do not edit by hand
\name{parameterReshape}
\alias{parameterReshape}
\title{parameterReshape}
\usage{
parameterReshape(paramList)
}
\arguments{
\item{paramList}{The parameter header to be reshaped}
}
\description{
Description:
	Used to convert lists of lists into matrices inorder to present
	the data in a more compact easier to read, and access, way.
}

