% Generated by roxygen2 (4.0.1): do not edit by hand
\name{readFile}
\alias{readFile}
\title{readFile}
\usage{
readFile(filename)
}
\arguments{
\item{filename}{input ODF file name}
}
\description{
Description:
  read the data from a file and return an array of lines
}
\details{
Output : OUTPUT = array of file data
}

