% Generated by roxygen2 (4.0.1): do not edit by hand
\name{writeUtf8}
\alias{writeUtf8}
\title{writeUtf8}
\usage{
writeUtf8(out, con)
}
\arguments{
\item{out}{String to be written to the file}

\item{con}{connection to the ODF file being written to}
}
\description{
Description:
  Used to write special windows characters like fancy quotes to the file
}

