% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odf.R
\name{od}
\alias{od}
\title{Create origin-destination flows (odf) object}
\usage{
od(
  od,
  points = NULL,
  col_orig = 1,
  col_dest = 2,
  col_flow = 3,
  col_type = NULL,
  remove_diag = FALSE
)
}
\arguments{
\item{od}{od}

\item{points}{points}

\item{col_orig}{col_orig}

\item{col_dest}{col_dest}

\item{col_flow}{col_flow}

\item{col_type}{col_type}

\item{remove_diag}{remove_diag}

\item{routes}{routes}

\item{col_via}{col_via}

\item{col_time}{col_time}
}
\description{
Create origin-destination flows (odf) object, which is a list consisting of a data.frame with the od data, an sf object with spatial points and optionally, an sf object with spatial (poly)lines of the routes
}
