% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odflows.R
\name{odf}
\alias{odf}
\title{Create origin-destination flows (odf) object}
\usage{
odf(od, points, routes = NULL, col_orig = 1, col_dest = 2,
  col_flow = 3, col_via = NULL, col_type = NULL, col_time = NULL)
}
\arguments{
\item{od}{od}

\item{points}{points}

\item{routes}{routes}

\item{col_orig}{col_orig}

\item{col_dest}{col_dest}

\item{col_flow}{col_flow}

\item{col_type}{col_type}

\item{col_time}{col_time}
}
\description{
Create origin-destination flows (odf) object, which is a list consisting of a data.frame with the od data, an sf object with spatial points and optionally, an sf object with spatial (poly)lines of the routes
}
