% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odf_map.R
\name{odf_flows}
\alias{odf_flows}
\alias{odf_points}
\title{Create map from odf object}
\usage{
odf_flows(x, by_type = TRUE, by_via = TRUE)

odf_points(odf)
}
\arguments{
\item{x}{odf object}

\item{by_type}{create flows for different type separately}

\item{by_via}{create flows for different via routes separately}
}
\description{
Create map from odf object
}
